fix: use config-inited
