+++ src/sphinx_book_theme/nodes.py
@@ -1,6 +1,7 @@
 from docutils import nodes
 from sphinx.application import Sphinx
 from typing import Any, cast
+from docutils.parsers.rst.directives.body import Sidebar
 
 
 class SideNoteNode(nodes.Element):
@@ -35,3 +36,22 @@ def depart_SideNoteNode(self, node):
     self.body.append(
         f"<input type='checkbox' id='{tagid}' name='{tagid}' class='margin-toggle'>"
     )
+
+
+class Margin(Sidebar):
+    """Goes in the margin to the right of the page."""
+
+    optional_arguments = 1
+    required_arguments = 0
+
+    def run(self):
+        """Run the directive."""
+        if not self.arguments:
+            self.arguments = [""]
+        nodes = super().run()
+        nodes[0].attributes["classes"].append("margin")
+
+        # Remove the "title" node if it is empty
+        if not self.arguments:
+            nodes[0].children.pop(0)
+        return nodes
